+++ tests/testthat/test-rank.R
@@ -141,6 +141,15 @@ test_that("`x` must be a vector", {
   expect_error(vec_rank(identity), class = "vctrs_error_scalar_type")
 })
 
+test_that("`x` must not be `NULL` (#1823)", {
+  expect_snapshot(error = TRUE, {
+    vec_rank(NULL)
+  })
+  expect_snapshot(error = TRUE, {
+    vec_rank(NULL, incomplete = "na")
+  })
+})
+
 test_that("`ties` is validated", {
   expect_snapshot(error = TRUE, vec_rank(1, ties = "foo"))
   expect_snapshot(error = TRUE, vec_rank(1, ties = 1))